+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file, and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Fixed bug in `Value`s tracking `set`s of numbers
+* Fixed bug in timestamp comparison in `BeforeAnnotation`
+* Added `Value.check_against` and `Attribute.check_agains` per [#111](https://github.com/microsoft/pybryt/issues/111)
+
 ## 0.2.0 2021-09-01
 
 * Fixed - 'Empty iterable comparison' per [#109](https://github.com/microsoft/pybryt/pull/109)